Make WinSparkle usable as a static library.
WinSparkle was written assuming it would be used as a DLL.  This change makes it possible to statically link WinSparkle into the main application or into another parent DLL.

This introduces a WINSPARKLE_STATIC #define that removes __declspec(dllimport) when statically linked.  It also removes the DllMain function, and replaces the call to SetDllHINSTANCE(hInstance) with an HINSTANCE lookup that will find the parent exe or the parent DLL, whichever applies.